+++ RELEASE.md
@@ -1,5 +1,7 @@
 # Upcoming release
 
+* Add a count data type to the data generator using a zero-inflated Poisson
+* Set bounds/max class imbalance for binary features for the data generators
 * Added Pytorch implementation for NOTEARS MLP which is much faster (Only supporting linear structure learning for now)
 * Added StructureRegressor sklearn interface using the Pytorch NOTEARS implementation.
 * Hotfix for data_gen system. Fixes issues with root node initialization.
