+++ tests/structure/data_generators/test_wrappers.py
@@ -26,28 +26,25 @@
 #
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import operator
 import string
 from itertools import product
-from typing import Hashable, Tuple, Union
 
-import networkx as nx
 import numpy as np
 import pandas as pd
 import pytest
-from networkx.algorithms.dag import is_directed_acyclic_graph
 from scipy.stats import anderson, stats
 
+from causalnex.structure import StructureModel
 from causalnex.structure.data_generators import (
     generate_binary_data,
     generate_binary_dataframe,
     generate_categorical_dataframe,
     generate_continuous_data,
     generate_continuous_dataframe,
+    generate_count_dataframe,
     generate_structure,
-    sem_generator,
 )
-from causalnex.structure.structuremodel import StructureModel
+from tests.structure.data_generators.test_core import calculate_proba
 
 
 @pytest.fixture
@@ -59,19 +56,6 @@ def graph():
 
 
 @pytest.fixture
-def schema():
-    # use the default schema for 3
-    schema = {
-        0: "binary",
-        1: "categorical:3",
-        2: "binary",
-        4: "continuous",
-        5: "categorical:5",
-    }
-    return schema
-
-
-@pytest.fixture()
 def graph_gen():
     def generator(num_nodes, seed, weight=None):
         np.random.seed(seed)
@@ -88,102 +72,6 @@ def graph_gen():
         return sm
 
     return generator
-
-
-class TestGenerateStructure:
-    @pytest.mark.parametrize("graph_type", ["erdos-renyi", "barabasi-albert", "full"])
-    def test_is_dag_graph_type(self, graph_type):
-        """ Tests that the generated graph is a dag for all graph types. """
-        degree, d_nodes = 4, 10
-        sm = generate_structure(d_nodes, degree, graph_type)
-        assert is_directed_acyclic_graph(sm)
-
-    @pytest.mark.parametrize("num_nodes,degree", [(5, 2), (10, 3), (15, 5)])
-    def test_is_dag_nodes_degrees(self, num_nodes, degree):
-        """ Tests that generated graph is dag for different numbers of nodes and degrees
-        """
-        sm = generate_structure(num_nodes, degree)
-        assert nx.is_directed_acyclic_graph(sm)
-
-    def test_bad_graph_type(self):
-        """ Test that a value other than "erdos-renyi", "barabasi-albert", "full" throws ValueError """
-        graph_type = "invalid"
-        degree, d_nodes = 4, 10
-        with pytest.raises(ValueError, match="unknown graph type"):
-            generate_structure(d_nodes, degree, graph_type)
-
-    @pytest.mark.parametrize("num_nodes,degree", [(5, 2), (10, 3), (15, 5)])
-    def test_expected_num_nodes(self, num_nodes, degree):
-        """ Test that generated structure has expected number of nodes = num_nodes """
-        sm = generate_structure(num_nodes, degree)
-        assert len(sm.nodes) == num_nodes
-
-    @pytest.mark.parametrize(
-        "num_nodes,degree,w_range",
-        [(5, 2, (1, 2)), (10, 3, (100, 200)), (15, 5, (1.0, 1.0))],
-    )
-    def test_weight_range(self, num_nodes, degree, w_range):
-        """ Test that w_range is respected in output """
-        w_min = w_range[0]
-        w_max = w_range[1]
-        sm = generate_structure(num_nodes, degree, w_min=w_min, w_max=w_max)
-        assert all(abs(sm[u][v]["weight"]) >= w_min for u, v in sm.edges)
-        assert all(abs(sm[u][v]["weight"]) <= w_max for u, v in sm.edges)
-
-    @pytest.mark.parametrize("num_nodes", [-1, 0, 1])
-    def test_num_nodes_exception(self, num_nodes):
-        """ Check a single node graph can't be generated """
-        with pytest.raises(ValueError, match="DAG must have at least 2 nodes"):
-            generate_structure(num_nodes, 1)
-
-    def test_min_max_weights_exception(self):
-        """ Check that w_range is valid """
-        with pytest.raises(
-            ValueError,
-            match="Absolute minimum weight must be less than or equal to maximum weight",
-        ):
-            generate_structure(4, 1, w_min=0.5, w_max=0)
-
-    def test_min_max_weights_equal(self):
-        """ If w_range (w, w) has w=w, check abs value of all weights respect this """
-        w = 1
-        sm = generate_structure(4, 1, w_min=w, w_max=w)
-        w_mat = nx.to_numpy_array(sm)
-        assert np.all((w_mat == 0) | (w_mat == w) | (w_mat == -w))
-
-    def test_erdos_renyi_degree_increases_edges(self):
-        """ Erdos-Renyi degree increases edges """
-        edge_counts = [
-            max(
-                [
-                    len(generate_structure(100, degree, "erdos-renyi").edges)
-                    for _ in range(10)
-                ]
-            )
-            for degree in [10, 90]
-        ]
-
-        assert edge_counts == sorted(edge_counts)
-
-    def test_barabasi_albert_degree_increases_edges(self):
-        """ Barabasi-Albert degree increases edges """
-        edge_counts = [
-            max(
-                [
-                    len(generate_structure(100, degree, "barabasi-albert").edges)
-                    for _ in range(10)
-                ]
-            )
-            for degree in [10, 90]
-        ]
-
-        assert edge_counts == sorted(edge_counts)
-
-    def test_full_network(self):
-        """ Fully connected network has expected edge counts """
-        sm = generate_structure(40, degree=0, graph_type="full")
-
-        assert len(sm.edges) == (40 * 39) / 2
 
 
 class TestGenerateContinuousData:
@@ -765,240 +653,12 @@ class TestGenerateCategoricalData:
                 assert np.isclose(joint_proba, factored_proba, rtol=tol, atol=0)
 
 
-class TestMixedDataGen:
-    def test_run(self, graph, schema):
-        df = sem_generator(
-            graph=graph,
-            schema=schema,
-            default_type="continuous",
-            noise_std=1.0,
-            n_samples=1000,
-            intercept=False,
-            seed=12,
-        )
-
-        # test binary:
-        assert df[0].nunique() == 2
-        assert df[0].nunique() == 2
-
-        # test categorical:
-        for col in ["1_{}".format(i) for i in range(3)]:
-            assert df[col].nunique() == 2
-        assert (
-            len([x for x in list(df.columns) if isinstance(x, str) and "1_" in x]) == 3
-        )
-
-        for col in ["5_{}".format(i) for i in range(5)]:
-            assert df[col].nunique() == 2
-        assert (
-            len([x for x in list(df.columns) if isinstance(x, str) and "5_" in x]) == 5
-        )
-
-        # test continuous
-        assert df[3].nunique() == 1000
-        assert df[4].nunique() == 1000
-
-    def test_graph_not_a_dag(self):
-        graph = StructureModel()
-        graph.add_edges_from([(0, 1), (1, 2), (2, 0)])
-
-        with pytest.raises(ValueError, match="Provided graph is not a DAG"):
-            _ = sem_generator(graph=graph)
-
-    def test_not_permissible_type(self, graph):
-        schema = {
-            0: "unknown data type",
-        }
-        with pytest.raises(ValueError, match="Unknown data type"):
-            _ = sem_generator(
-                graph=graph,
-                schema=schema,
-                default_type="continuous",
-                noise_std=1.0,
-                n_samples=1000,
-                intercept=False,
-                seed=12,
-            )
-
-    def test_missing_cardinality(self, graph):
-        schema = {
-            0: "categorical",
-            1: "categorical:3",
-            5: "categorical:5",
-        }
-        with pytest.raises(ValueError, match="Missing cardinality for categorical"):
-            _ = sem_generator(
-                graph=graph,
-                schema=schema,
-                default_type="continuous",
-                noise_std=1.0,
-                n_samples=1000,
-                intercept=False,
-                seed=12,
-            )
-
-    def test_missing_default_type(self, graph):
-        with pytest.raises(ValueError, match="Unknown default data type"):
-            _ = sem_generator(
-                graph=graph,
-                schema=schema,
-                default_type="unknown",
-                noise_std=1.0,
-                n_samples=1000,
-                intercept=False,
-                seed=12,
-            )
-
-    def test_incorrect_weight_dist(self):
-        sm = StructureModel()
-        nodes = list(str(x) for x in range(6))
-        np.random.shuffle(nodes)
-        sm.add_nodes_from(nodes)
-
-        sm.add_weighted_edges_from([("0", "1", None), ("2", "4", None)])
-
-        with pytest.raises(ValueError, match="Unknown weight distribution"):
-            _ = sem_generator(
-                graph=sm,
-                schema=None,
-                default_type="continuous",
-                distributions={"weight": "unknown"},
-                noise_std=2.0,
-                n_samples=1000,
-                intercept=False,
-                seed=10,
-            )
-
-    def test_incorrect_intercept_dist(self, graph):
-        with pytest.raises(ValueError, match="Unknown intercept distribution"):
-            _ = sem_generator(
-                graph=graph,
-                schema=None,
-                default_type="continuous",
-                distributions={"intercept": "unknown"},
-                noise_std=2.0,
-                n_samples=10,
-                intercept=True,
-                seed=10,
-            )
-
-    # Seed 20 is an unlucky seed and fails the assertion. All other seeds tested
-    # pass the assertion. Similar issue to the categorical intercept test?
-    @pytest.mark.parametrize("seed", (10, 17))
-    @pytest.mark.parametrize("n_categories", (2, 5,))
-    @pytest.mark.parametrize("weight_distribution", ["uniform", "gaussian"])
-    @pytest.mark.parametrize("intercept_distribution", ["uniform", "gaussian"])
-    def test_mixed_type_independence(
-        self, seed, n_categories, weight_distribution, intercept_distribution
-    ):
-        """
-        Test whether the relation is accurate, implicitly tests sequence of
-        nodes.
-        """
-        np.random.seed(seed)
-
-        sm = StructureModel()
-        nodes = list(str(x) for x in range(6))
-        np.random.shuffle(nodes)
-        sm.add_nodes_from(nodes)
-        # binary -> categorical
-        sm.add_weighted_edges_from([("0", "1", 10)])
-        # binary -> continuous
-        sm.add_weighted_edges_from([("2", "4", None)])
-
-        schema = {
-            "0": "binary",
-            "1": "categorical:{}".format(n_categories),
-            "2": "binary",
-            "4": "continuous",
-            "5": "categorical:{}".format(n_categories),
-        }
-
-        df = sem_generator(
-            graph=sm,
-            schema=schema,
-            default_type="continuous",
-            distributions={
-                "weight": weight_distribution,
-                "intercept": intercept_distribution,
-            },
-            noise_std=2,
-            n_samples=100000,
-            intercept=True,
-            seed=seed,
-        )
-
-        atol = 0.05  # 5% difference bewteen joint & factored!
-        # 1. dependent links
-        # 0 -> 1 (we look at the class with the highest deviation from uniform
-        # to avoid small values)
-        c, _ = max(
-            [
-                (c, np.abs(df["1_{}".format(c)].mean() - 1 / n_categories))
-                for c in range(n_categories)
-            ],
-            key=operator.itemgetter(1),
-        )
-        joint_proba, factored_proba = calculate_proba(df, "0", "1_{}".format(c))
-        assert not np.isclose(joint_proba, factored_proba, rtol=0, atol=atol)
-        # 2 -> 4
-        assert not np.isclose(
-            df["4"].mean(), df["4"][df["2"] == 1].mean(), rtol=0, atol=atol
-        )
-
-        tol = 0.15  # relative tolerance of +- 15% of the
-        # 2. independent links
-        # categorical
-        c, _ = max(
-            [
-                (c, np.abs(df["1_{}".format(c)].mean() - 1 / n_categories))
-                for c in range(n_categories)
-            ],
-            key=operator.itemgetter(1),
-        )
-        joint_proba, factored_proba = calculate_proba(df, "0", "5_{}".format(c))
-        assert np.isclose(joint_proba, factored_proba, rtol=tol, atol=0)
-
-        # binary
-        joint_proba, factored_proba = calculate_proba(df, "0", "2")
-        assert np.isclose(joint_proba, factored_proba, rtol=tol, atol=0)
-
-        # categorical
-        c, _ = max(
-            [
-                (c, np.abs(df["1_{}".format(c)].mean() - 1 / n_categories))
-                for c in range(n_categories)
-            ],
-            key=operator.itemgetter(1),
-        )
-        d, _ = max(
-            [
-                (d, np.abs(df["5_{}".format(d)].mean() - 1 / n_categories))
-                for d in range(n_categories)
-            ],
-            key=operator.itemgetter(1),
-        )
-        joint_proba, factored_proba = calculate_proba(
-            df, "1_{}".format(d), "5_{}".format(c)
-        )
-        assert np.isclose(joint_proba, factored_proba, rtol=tol, atol=0)
-
-        # continuous
-        # for gaussian distributions, zero variance is equivalent to independence
-        assert np.isclose(df[["3", "4"]].corr().values[0, 1], 0, atol=tol)
-
-
-def calculate_proba(
-    df: Union[pd.DataFrame, np.ndarray], col_0: Hashable, col_1: Hashable
-) -> Tuple[float, float]:
-    if isinstance(df, pd.DataFrame):
-        marginal_0 = df[col_0].mean()
-        marginal_1 = df[col_1].mean()
-        joint_proba = (df[col_0] * df[col_1]).mean()
-    else:
-        marginal_0 = df[:, col_0].mean()
-        marginal_1 = df[:, col_1].mean()
-        joint_proba = (df[:, col_0] * df[:, col_1]).mean()
-
-    factored_proba = marginal_0 * marginal_1
-    return joint_proba, factored_proba
+def test_zero_lambda():
+    """
+    A wrong initialisation could lead to counts always being zero if they dont
+    have parents.
+    """
+    graph = StructureModel()
+    graph.add_nodes_from(list(range(20)))
+    df = generate_count_dataframe(graph, 10000)
+    assert not np.any(df.mean() == 0)
